Fix WIN32 preprocessor directives to use _WIN32
The WIN32 macro has been deprecated, and is no longer supported in the
most recent versions of MSYS2. In order to facilitate continued support
for MSYS2, this change replaces all instances of WIN32 with _WIN32 the
macro which is recommended for Windows platform detection.

This commit also fixes an issue where `#ifndef isfinite` was no longer
detecting the windows system builtin isfinite() and was redefining the
function to the linux platform variant finite().