Add buffer to prevent runaway allocations
The unordered map allocations were showing problematic allocation
properties, and a sized buffer and cycle count was added such
that long-transition waterheaters will not continue to allocate
when the old data is not used. to avoid seek collisions, a buffer
size of 256 was used, and detections of no-change time to
transitions exceeding this will purge their buffers and reuse the
memory, to avoid continuous growth.