Add buffer control variable to waterheater
This commit adds a new variable called state_change_buffer_size
which can be optionally used to increase the internal buffers
used by the state change tracking component of the time-to-transition
logic. The default value is set to 256 double precision values, but
can be increased arbitrarily.
Larger buffers will meaningfully increase the memory footprint of
the simulation, but may reduce the cycle load of the memory clear
in simulations with significant time-to-transition overhead.