First sketch of controller EOD commit
First sketch of secondary controller in inter_deltaupdate.
Idea right now is that there will *not* be any predictor corrector
but rather just passthrough behavior.
However, this leaves the question open for how to handle the outputs
so we don't accumulate things unintentionally during iteration.

For example, say a generator get's a deltaP setpoint change on
iteration 0. Then the second iteration is performed, leading to a
different deltaP.
- Is it correct/what we want to simply add these two deltas effectively
in one timestep?
    * Doing this actually leads to seeming incorrect controller behavior

- one possibility is to only publish the values following all
iterations...That's probably the more sensible action.