Fix Pref/Qref initialization in P_CONSTANT mode
Fixes #1345.

P_CONSTANT mode and SEXS_CQ mode use Pref and Qref to set the
output power. If these are not specified in the glm they have
the value of -99.0 pu in the init call, which is clearly erroneous.
This fix sets Pref and Qref to the sum of power_out_X
(real and imaginary repsectively) divided by the rated VA and issues
a warning that in general Pref/Qref should be specified.