Allow diesel_dg to initialize to zero
Closes #1356

Previously, if a diesel_dg generator was initialized with any of the
power_out_X properties set to 0, a warning was raised and the power
for that phase set to 50%.
This meant that it was not possible to initialize a generator as
"off line".

This fixes this issue, by removing the arbitrary change to 50%.
The warning is kept as a precaution, since it is only in very particular
cases, where it truely makes sense to initialize a generator to zero.