Include unit error in error signal (NOT TESTED!)
The difference between the sum of all Pset - Pout is now subtracted
from the frequency error (multiplied by a bias factor).
This effectively says, that we consider the time delay in generation
reaching its set value.

For example, our frequency error says we need to add more generation,
but we see that the generation produced is below the set references.
This means that the generation will continue to increase until it
reaches the desired set points. We should therefore, remove the
difference between set point and actual output from the frequency error
signal.