Trip status check in inverter_dyn pdispatch
When synchronizing the pdispatch variables with the actual controller
setpoints in the function pdispatch_synch the IEEE 1547 trip status
is now considered.

If the device is tripped the pdispatch and the offset are set to 0.
If changes are made to pdispatch or the offset these are OVERWRITTEN
and the variables remain 0.
Note that the actuall reference variables are NOT altered.
Therefore, when the trip status changes, the last reference setpoint
will still be present, and pdispatch will be altered accordingly
(This could mean that there is a different between pdispatch and
pdispatch offset before and after the trip. Example:
Before trip: pdispatch = 0.5, pdispatch_offset = 0.1 -> Pref = 0.6*Sbase
After trip: Pref = 0.6*Spbase -> pdispatch = 0.6, pdispatch_offset = 0.
)

Background:
The reason for this change is that the when the unit was tripping, the
secondary controller was still sending it change signals. As a result
the unit error was increasing and offsetting the frequency error,
leading the controller to settle at a non-nominal frequency.

A "better" solution will likely somehow communicate the trip to the
secondary controller so that the participation is re-allocated.
However, the simple fix described here still work for reestablishing
nominal frequency, and a tertiary control level could then take charge
to take note of the trip and reallocate things in the secondary level.