Improving low pass filter to participants
The implementation of the low pass filter to participants was erroneous
in that it assument that when Tlp = dt the low pass filter would become
a pass through. In fact, the result was 1/2 of the value to pass through
which actually agrees with implicit integration approximation.

The default Tlp was changed to 0 which is now handled slightely
differently to allow the pass through.

The hand off to the participating objects was also altered since
previously it was only taking place on sample instances, which defeats
the purpose of a low pass filter.

The sought after behavior is described by the following table:
|Sample Change | Tlp > 0 | behavior                                    |
|--------------|---------|---------------------------------------------|
| Yes          | Yes     | PID output modifies current setpoint        |
|              |         | Pass to participants via low pass integrator|
|--------------|---------|---------------------------------------------|
| Yes          | No      | PID output modifies current setpoint        |
|              |         | Pass value directly to participatns         |
|--------------|---------|---------------------------------------------|
| No           | Yes     | Participants dP variable slowly converges to|
|              |         | Value from last sample instant              |
|              |         | Only the incremental change should be passed|
|              |         | to participatns                             |
|--------------|---------|---------------------------------------------|
| No           | No      | No update should take place                 |