Resolve Windows pathing issues and disable Mac ARM64 build
the ARM64 instructions do not appear to be working at present,
and as such I've disabled them until there is better support
on the Github actions official workflow.

We may need to publish a specific build workflow for users of
new Mac systems so that they can build for their architecture.