Adding sampling to inputs
Unit error and frequency error inputs are now sampled at rate Ts_P
and Ts_f respectively. This increases the flexibility and allows
to model slower communication rates.

Note that while the default for Ts_f is deltat (gridlab-d's timestep),
the default for Ts_p is Ts, the output sampling rate.

The PID only updates if either Ts_P or Ts_f is up. For this reason, the
integration step is no longer gridlab's deltat but min(Ts_P, Ts_f).

The alpha_check function was also removed because it was not being used
nor was use intended. The function was there to make sure the sum of all
alphas summed to one, but
a. there is no real reason to enforce this
b. this is not true the way alpha has been overloaded for tie-lines
(although that could be excepted)

At this commit the code compiles successfully but has not yet been
tested.