Fix to int_assert
Removed test for 32 bit maximum size.
The issue was that both compared quantities (x and value) were converted
to int64 before this test.
As a result any test that correctly checks that the difference of the two
is within the range of 32bit integers would necessarily pass, because
the coversion has already taken place.

A different issue is that the test was implemented incorrectly and
theDiff < -0xFFFFFFFF passed whenever theDiff=0, which is obviously not
the desired behavior.