Changes to secondary controller in sync
Added a call to parse_participant_input in the bottom-up pass of
the sync call.
This way, although the secondary controller doesn't do anything in
Event Mode, it can still be updated without having to enter Deltamode.

Possible example:
1. Secondary controller is parameterized in some fashion
2. Event occurs and the secondary controller reacts accordingly
3. Steady state is restablished and simulation goes back to QSTS***
4. The secondary controller is re-parametrized in anticipation of some
future event.
5. Another event is simulated in Deltamode.

*** Note that this is a bit tricky at the moment since rotor convergence
has to be set very small to ensure there is not drift following a return
to QSTS. As a result, it is not that easy to actually exit Deltamode.