Creating Autotests for the secondary controller
Test 1 goes through a load step event and verifies that the dispatch
signal to the participating generators as well as their actuall power
output behaves as expected.
(Note: if/when reactive capabilities will be added this response might
change)

Test 2 simply checks that objects can be correctly added, removed, and
modified in the secondary controller.