Update module bindings to use fixed parameters instead of variadics
The ARM64 architecture interacts with variadic functions in a
different way than x86 architecture, resulting in attempted memory
accesses outside of allocated scope when trying to enumerate the
parameters in the FUNCTIONADDR binding.

To resolve this issue, each control-plane method used to interface
with the GridLAB-D modules has been replaced with a fully
type-defined handle, such that the variadic enumeration issues
across the module boundry are no longer present.